requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Image processing
numpy>=1.24.0
pillow>=10.0.0
scipy>=1.11.0
scikit-image>=0.22.0

# Cell segmentation (Cellpose only — StarDist disabled)
cellpose>=3.0.0,<4.0.0

# Monitoring
psutil>=5.9.0
